requirements.txt
aiohttp==3.9.5
anyio==4.6.2.post1
beanie==1.27.0
beautifulsoup4==4.12.3
bilireq==0.2.13
django_recaptcha==4.0.0
dynamicadaptor==0.5.5
dynrender_skia==0.2.9
grpcio==1.67.1
httpx==0.27.2
Jinja2==3.1.4
loguru==0.7.2
lxml==5.3.0
nonebot==1.9.1
nonebot_plugin_htmlrender==0.4.0
openpyxl==3.1.5
optional==0.0.1
Pillow==11.0.0
playwright==1.49.0
protobuf==5.28.3
pydantic==2.10.1
pynng==0.8.0
pypinyin==0.53.0
pyspellchecker==0.8.1
pytz==2024.2
PyYAML==6.0.2
PyYAML==6.0.2
qrcode==8.0
Requests==2.32.3
schedule==1.2.2
subscribe==0.6.3
tiktoken==0.8.0
typing_extensions==4.12.2
ujson==5.10.0
wcferry==39.2.4.0
xmltodict==0.13.0
yarl==1.17.1
